@@ -3,13 +3,14 @@ package com.livelike.engagementsdk.widget.services.messaging.pubnub
 import com.pubnub.api.PubNub
 import com.pubnub.api.callbacks.SubscribeCallback
 import com.pubnub.api.models.consumer.PNStatus
+import com.pubnub.api.models.consumer.objects_api.channel.PNChannelMetadataResult
+import com.pubnub.api.models.consumer.objects_api.membership.PNMembershipResult
+import com.pubnub.api.models.consumer.objects_api.uuid.PNUUIDMetadataResult
 import com.pubnub.api.models.consumer.pubsub.PNMessageResult
 import com.pubnub.api.models.consumer.pubsub.PNPresenceEventResult
 import com.pubnub.api.models.consumer.pubsub.PNSignalResult
+import com.pubnub.api.models.consumer.pubsub.files.PNFileEventResult
 import com.pubnub.api.models.consumer.pubsub.message_actions.PNMessageActionResult
-import com.pubnub.api.models.consumer.pubsub.objects.PNMembershipResult
-import com.pubnub.api.models.consumer.pubsub.objects.PNSpaceResult
-import com.pubnub.api.models.consumer.pubsub.objects.PNUserResult
 
 /**
  * This adapter class provides empty implementations of the methods from pubnub subscribe callbacks.
@@ -24,8 +25,6 @@ internal abstract class PubnubSubscribeCallbackAdapter : SubscribeCallback() {
     override fun status(pubnub: PubNub, pnStatus: PNStatus) {
     }
 
-    override fun user(pubnub: PubNub, pnUserResult: PNUserResult) {
-    }
 
     override fun messageAction(pubnub: PubNub, pnMessageActionResult: PNMessageActionResult) {
     }
@@ -39,6 +38,13 @@ internal abstract class PubnubSubscribeCallbackAdapter : SubscribeCallback() {
     override fun message(pubnub: PubNub, pnMessageResult: PNMessageResult) {
     }
 
-    override fun space(pubnub: PubNub, pnSpaceResult: PNSpaceResult) {
+    override fun uuid(pubnub: PubNub, pnUUIDMetadataResult: PNUUIDMetadataResult) {
     }
+
+    override fun file(pubnub: PubNub, pnFileEventResult: PNFileEventResult) {
+    }
+
+    override fun channel(pubnub: PubNub, pnChannelMetadataResult: PNChannelMetadataResult) {
+    }
+
 }
